fix app.jsx
